requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
Click==7.0
decorator==4.4.1
Django==3.0.2
django-webpack-loader==0.6.0
djangorestframework==3.11.0
dodgy==0.2.1
gunicorn==20.0.4
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-bakery==1.1.0
parso==0.6.0
pathspec==0.7.0
pep8-naming==0.4.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
prospector==1.2.0
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.4.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requirements-detector==0.6
setoptconf==0.2.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.2
